initial version
web page that uses javascript to create JSON object that machData.py uses as relational data model definition. Mock data is created following the relationships described by this JSON object.  
This is not a fully working versoin ( does not for example download a working version of the mach data JSON object. It is a work in progress